Struct trillium_http::Stopper

pub struct Stopper(_);

This struct provides a synchronized mechanism for canceling Futures and Streams.

Implementations

impl Stopper

pub fn new() -> Stopper

Initialize a stopper that is not yet stopped and that has zero registered wakers. Any clone of this stopper represents the same internal state. This is identical to Stopper::default()

pub fn stop(&self)

Stop all futures and streams that have been registered to this Stopper or any clone representing the same initial stopper.

pub fn is_stopped(&self) -> bool

Returns whether this stopper (or any clone of it) has been stopped.

Example

let stopper = stopper::Stopper::new();
assert!(!stopper.is_stopped());
stopper.stop();
assert!(stopper.is_stopped());

pub fn stop_stream<S>(&self, stream: S) -> StreamStopper<S>where S: Stream,

This function returns a new stream which will poll None (indicating a completed stream) when this Stopper has been stopped. The Stream’s Item is unchanged.

Example

use futures_lite::StreamExt;
let stopper = stopper::Stopper::new();
let mut stream = stopper.stop_stream(futures_lite::stream::repeat("infinite stream"));

std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    stopper.stop();
});

while let Some(item) = stream.next().await {
    println!("{}", item);
}

pub fn stop_future<F>(&self, future: F) -> FutureStopper<F>where F: Future,

This function returns a Future which wraps the provided future and stops it when this Stopper has been stopped. Note that the Output of the returned future is wrapped with an Option. If the future resolves to None, that indicates that it was stopped instead of polling to completion.

Example

let stopper = stopper::Stopper::new();
let mut future = stopper.stop_future(std::future::pending::<()>());

std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    stopper.stop();
});

assert_eq!(future.await, None);