Struct trillium_http::transport::BoxedTransport

pub struct BoxedTransport(_);

A type for dyn Transport trait objects

BoxedTransport represents a Box<dyn Transport> that supports downcasting to the original Transport. This is used in trillium to erase the generic on Conn, in order to avoid writing Conn<TcpStream> throughout an application.

Stability Note: This may go away at some point and be replaced by a type alias exported from a trillium_server_common::Server crate.

Implementations

impl BoxedTransport

pub fn new(t: impl Transport + 'static) -> Self

Create a new BoxedTransport from some Transport.

use trillium_http::transport::BoxedTransport;
use trillium_testing::TestTransport;
// this examples uses trillium_testing::TestTransport as an
// easily-constructed AsyncRead+AsyncWrite.
let (test_transport, _) = TestTransport::new();

let boxed = BoxedTransport::new(test_transport);

let downcast: Option<Box<TestTransport>> = boxed.downcast();
assert!(downcast.is_some());

let boxed_again = BoxedTransport::new(*downcast.unwrap());
assert!(boxed_again.downcast::<async_net::TcpStream>().is_none());

pub fn downcast<T: 'static>(self) -> Option<Box<T>>

Attempt to convert the trait object into a specific transport T. This will only succeed if T is the type that was originally passed to BoxedTransport::new, and will return None otherwise

see BoxedTransport::new for example usage