Crate trillium_http

This crate provides the http 1.x implementation for Trillium.

Stability

As this is primarily intended for internal use by the Trillium crate, the api is likely to be less stable than that of the higher level abstractions in Trillium.

Example

This is an elaborate example that demonstrates some of trillium_http’s capabilities. Please note that trillium itself provides a much more usable interface on top of trillium_http, at very little cost.

use async_net::{TcpListener, TcpStream};
use futures_lite::StreamExt;
use stopper::Stopper;
use trillium_http::{Conn, Result};

let stopper = Stopper::new();
let listener = TcpListener::bind(("localhost", 0)).await?;
let port = listener.local_addr()?.port();

let server_stopper = stopper.clone();
let server_handle = smol::spawn(async move {
    let mut incoming = server_stopper.stop_stream(listener.incoming());

    while let Some(Ok(stream)) = incoming.next().await {
        let stopper = server_stopper.clone();
        smol::spawn(Conn::map(stream, stopper, |mut conn: Conn<TcpStream>| async move {
            conn.set_response_body("hello world");
            conn.set_status(200);
            conn
         })).detach()
    }

    Result::Ok(())
});

// this example uses the trillium client
// please note that this api is still especially unstable.
// any other http client would work here too
let url = format!("http://localhost:{}/", port);
type ClientConn<'a> = trillium_client::Conn<'a, trillium_smol::TcpConnector>;
let mut client_conn = ClientConn::get(&*url).execute().await?;

assert_eq!(client_conn.status().unwrap(), 200);
assert_eq!(client_conn.response_headers()["content-length"], "11");
assert_eq!(
    client_conn.response_body().read_string().await?,
    "hello world"
);

stopper.stop(); // stop the server after one request
server_handle.await?; // wait for the server to shut down

Re-exports

pub use http_types;

Modules

transport
Types to represent the bidirectional data stream over which the HTTP protocol is communicated

Structs

ChunkedEncoder
An encoder for chunked encoding.

Conn
A http connection

ReceivedBody
A received http body

Stopper
This struct provides a synchronized mechanism for canceling Futures and Streams.

Synthetic
Synthetic represents a simple transport that contains fixed content. This is exclusively useful for testing or for server implementations that are not read from an io connection, such as a faas function, in which the entire body may be available immediately on invocation.

Upgrade
This open (pub fields) struct represents a http upgrade. It contains all of the data available on a Conn, as well as owning the underlying transport.

Enums

BodyEncoder
A http encoder for [http_types::Body]. You probably don’t want to interact with this directly.

ConnectionStatus
This represents the next state after a response on a conn transport.

Error
Concrete errors that occur within trillium’s http implementation

ReceivedBodyState
the current read state of this body

Type Definitions

Result
this crate’s result type