Struct trillium_http::ReceivedBody

pub struct ReceivedBody<'conn, Transport> { /* private fields */ }

A received http body

This type represents a body that will be read from the underlying transport, which it may either borrow from a Conn or own.

let mut conn = Conn::new_synthetic(Method::Get, "/", "hello");
let body = conn.request_body().await;
assert_eq!(body.read_string().await?, "hello");

Implementations

impl<'conn, Transport> ReceivedBody<'conn, Transport> where
Transport: AsyncRead + Unpin + Send + Sync + 'static,

pub fn content_length(&self) -> Option<u64>

Returns the content-length of this body, if available. This usually is derived from the content-length header. If the http request or response that this body is attached to uses transfer-encoding chunked, this will be None.

let mut conn = Conn::new_synthetic(Method::Get, "/", "hello");
let body = conn.request_body().await;
assert_eq!(body.content_length(), Some(5));

pub async fn read_string(self) -> Result<String>

Reads the entire body to string, using the encoding determined by the content-type (mime) charset. If an encoding problem is encountered, the String returned by read_string will contain utf8 replacement characters.

Note that this can only be performed once per Conn, as the underlying data is not cached anywhere. This is the only copy of the body contents.

Errors

This will return an error if there is an IO error on the underlying transport such as a disconnect

pub async fn read_bytes(self) -> Result<Vec<u8>>

Similar to ReceivedBody::read_string, but returns the raw bytes. This is useful for bodies that are not text.

You can use this in conjunction with encoding if you need different handling of malformed character encoding than the lossy conversion provided by read_string.

Errors

This will return an error if there is an IO error on the underlying transport such as a disconnect

pub fn encoding(&self) -> &'static Encoding

returns the character encoding of this body, usually determined from the content type (mime-type) of the associated Conn.

pub async fn drain(self) -> Result<u64>

Consumes the remainder of this body from the underlying transport by reading it to the end and discarding the contents. This is important for http1.1 keepalive, but most of the time you do not need to directly call this. It returns the number of bytes consumed.

Errors

This will return an std::io::Result::Err if there is an io error on the underlying transport, such as a disconnect

impl<T> ReceivedBody<'static, T>

pub fn take_transport(&mut self) -> Option<T>

takes the static transport from this received body