Struct trillium_http::ReceivedBody

pub struct ReceivedBody<'conn, Transport> { /* private fields */ }

A received http body

This type represents a body that will be read from the underlying transport, which it may either borrow from a Conn or own.

let mut conn = Conn::new_synthetic(Method::Get, "/", "hello");
let body = conn.request_body().await;
assert_eq!(body.read_string().await?, "hello");

Bounds checking

Every ReceivedBody has a maximum length beyond which it will return an error, expressed as a u64. To override this on the specific ReceivedBody, use ReceivedBody::with_max_len or ReceivedBody::set_max_len

The default maximum length is currently set to 500mb. In the next semver-minor release, this value will decrease substantially.

Large chunks, small read buffers

Attempting to read a chunked body with a buffer that is shorter than the chunk size in hex will result in an error. This limitation is temporary.

Implementations

impl<'conn, Transport> ReceivedBody<'conn, Transport>
where Transport: AsyncRead + Unpin + Send + Sync + 'static,

pub fn content_length(&self) -> Option<u64>

Returns the content-length of this body, if available. This usually is derived from the content-length header. If the http request or response that this body is attached to uses transfer-encoding chunked, this will be None.

let mut conn = Conn::new_synthetic(Method::Get, "/", "hello");
let body = conn.request_body().await;
assert_eq!(body.content_length(), Some(5));

pub async fn read_string(self) -> Result<String>

Reads entire body to String.

This uses the encoding determined by the content-type (mime) charset. If an encoding problem is encountered, the String returned by ReceivedBody::read_string will contain utf8 replacement characters.

Note that this can only be performed once per Conn, as the underlying data is not cached anywhere. This is the only copy of the body contents.

Errors

This will return an error if there is an IO error on the underlying transport such as a disconnect

This will also return an error if the length exceeds the maximum length. To override this value on this specific body, use ReceivedBody::with_max_len or ReceivedBody::set_max_len

pub fn set_max_len(&mut self, max_len: u64)

Set the maximum length that can be read from this body before error

See also HttpConfig::received_body_max_len

pub fn with_max_len(self, max_len: u64) -> Self

chainable setter for the maximum length that can be read from this body before error

See also HttpConfig::received_body_max_len

pub async fn read_bytes(self) -> Result<Vec<u8>>

Similar to ReceivedBody::read_string, but returns the raw bytes. This is useful for bodies that are not text.

You can use this in conjunction with encoding if you need different handling of malformed character encoding than the lossy conversion provided by ReceivedBody::read_string.

Errors

This will return an error if there is an IO error on the underlying transport such as a disconnect

This will also return an error if the length exceeds received_body_max_len. To override this value on this specific body, use ReceivedBody::with_max_len or ReceivedBody::set_max_len

pub fn encoding(&self) -> &'static Encoding

returns the character encoding of this body, usually determined from the content type (mime-type) of the associated Conn.

pub async fn drain(self) -> Result<u64>

Consumes the remainder of this body from the underlying transport by reading it to the end and discarding the contents. This is important for http1.1 keepalive, but most of the time you do not need to directly call this. It returns the number of bytes consumed.

Errors

This will return an std::io::Result::Err if there is an io error on the underlying transport, such as a disconnect

impl<T> ReceivedBody<'static, T>

pub fn take_transport(&mut self) -> Option<T>

takes the static transport from this received body