Struct trillium_http::Upgrade

pub struct Upgrade<Transport> {
    pub request_headers: Headers,
    pub path: String,
    pub method: Method,
    pub state: StateSet,
    pub transport: Transport,
    pub buffer: Option<Vec<u8>>,
    pub stopper: Stopper,
}

This open (pub fields) struct represents a http upgrade. It contains all of the data available on a Conn, as well as owning the underlying transport.

Important implementation note: When reading directly from the transport, ensure that you read from buffer first if there are bytes in it. Alternatively, read directly from the Upgrade, as that [AsyncRead] implementation will drain the buffer first before reading from the transport.

Fields
request_headers: Headers

The http request headers

path: String

The request path

method: Method

The http request method

state: StateSet

Any state that has been accumulated on the Conn before negotiating the upgrade

transport: Transport

The underlying io (often a TcpStream or similar)

buffer: Option<Vec<u8>>

Any bytes that have been read from the underlying tcpstream already. It is your responsibility to process these bytes before reading directly from the transport.

stopper: Stopper

A Stopper which can and should be used to gracefully shut down any long running streams or futures associated with this upgrade

Implementations

impl<Transport> Upgrade<Transport>

pub fn headers(&self) -> &Headers

read-only access to the request headers

pub fn path(&self) -> &str

the http request path up to but excluding any query component

pub fn querystring(&self) -> &str

retrieves the query component of the path

pub fn method(&self) -> &Method

the http method

pub fn state(&self) -> &StateSet

any state that has been accumulated on the Conn before negotiating the upgrade.

pub fn map_transport<T: AsyncRead + AsyncWrite + Send + Sync + Unpin + 'static>( self, f: impl Fn(Transport) -> T ) -> Upgrade<T>

Modify the transport type of this upgrade. This is useful for boxing the transport in order to erase the type argument.